+++ reinforce/algo/__init__.py
@@ -1,2 +1,3 @@
-__all__ = []
+from rl_common.returns import discount_cumsum
 
+__all__ = ["discount_cumsum"]
