+++ reinforce/scripts/run_from_config.sh
